Record school move log entries
This updates the `confirm!` method on the `SchoolMove` model to record
an entry in the school move log when a school move is confirmed. This
acts in parallel with our existing audit logs that are based on database
changes, allowing us to track missing information such as whether the
user asked for the patient to remain in the community clinic or not.