Add recipient_deterministic column
This adds a new column to `NotifyLogEntries` which allows us to store a
deterministically encrypted version of the recipient on this table,
making it possible to query against the value. This first commit adds
the new column, which will then start to be used in subsequent commits.
Eventually, the old `recipient` column will be entirely replaced by this
one.