Handle consent label without a relationship
When displaying the label for a consent response, sometimes the
`parent_relationship` will be `nil`. This can happen if an ops task
manually removes the relationship, or if a patient is merged. Currently
when trying to view this patient we will see a server error whereas we
should instead display the name of the parent even if the relationship
no longer exists.

This will also be relevant in the future where we will give the ability
for nurses to remove parental contacts, but might still want a record of
the consent.