Add callback bearer tokens for GOV.UK Notify
This adds a randomly generated bearer token for staging and production
that we'll use to authenticate the delivery receipt callbacks from
GOV.UK Notify.

Only one bearer token/callback URL can be specified at any one time, so
I expect only the production token will be used in the future, but until
we've gone live with the feature, we'll configure the staging token in
GOV.UK Notify.