Only look for existing patients with an NHS number
When editing the NHS number of the patient and trying to remove the NHS
number, currently this ends up finding existing patients without NHS
numbers and flagging them as existing patients, which isn't the correct
behaviour. Instead, we should allow users to set the NHS number to a
blank string without trying to merge with an existing patient.