Add patient_date_of_birth personalisation
This adds a new personalisation variable that is sent up to GOV.UK
Notify when sending an email or text message, allowing the date of birth
to be used in any templates.

We're not going to update the templates yet but this allows us to
update them in the future without needing to make a code change at the
time.