Add foreign keys to school move log entries
This reverts the change that was made in
6d8880d3c5b4810b7851614fdf30ce365a163f6e to remove these foreign keys to
fix a bug where patients couldn't be merged due to the school move log
entries existing.

Instead, we can handle the school move log entries by modifying the
`patient_id` value, ensuring that the foreign keys continue to be valid.
Removing the foreign keys means we can no longer rely on this model to
generate activity logs which is one of the reason it was added.

It's true that it also prevents us from being able to delete any schools
or users, but we already have this problem or more crucial models (like
vaccination records, triage, sessions, etc) so when if we come to a
situation where we need to handle that we should do it in a consistent
way (likely by soft-deleting the foreign keyed model).