Increase wait time between PDS jobs
We're seeing a problem in production where the queues for PDS jobs gets
full with jobs that are not succeeding, likely related to the rate
limit.

Some potentially related issues:

- https://github.com/bensheldon/good_job/issues/838
- https://github.com/bensheldon/good_job/issues/1278
- https://github.com/bensheldon/good_job/issues/1353

This wait time was added to artifically slow down the queues and prevent
them from needing to be retried, so increasing this wait time should
reduce the changes of the jobs ending up in this stuck state.

This is a temporary measure until we can either move away from Good Job
or resolve the underlyling issue which is still not 100% clear.