Add migration to add patient and sessions to vaccination records
We need to decouple vaccination records from sessions, so they can exist
outside of a session. To do this we need to add references to the
patient and the location that we currently get from the patient session
reference.

Decoupling vaccination records from sessions is necessary to support
doubles where patients can be vaccinated outside of a session managed by
Mavis (for example in A&E or a pharmacy before travelling abroad). We
need to be able to import these records and allow them to be visible
across all organisations as approriate.