Filter triage and consents by programme
When handling the consents and triage we should filter by the current
programme, or the programmes for the current session. This is to avoid
problems where the incorrect consent, triage or vaccination record is
shown resulting in an incorrect status.

For the most part this hard codes the programme as the first programme
in the session and doesn't handle multiple programmes correctly, however
it does fix a few potential bugs where we can select a programme
approriately.

This has highlighted more locations that will break with multiple
programmes, but I anticipate they can all be fixed in a similar way once
we have finalised the designs.