Add performed_ods_code column
This adds a new column to vaccination records which tracks the ODS code
of the organisation that performed the vaccination (whether that's a
school immunisations team, a hospital or a GP). This is necessary to
support importing historical vaccinations that weren't performed by the
SAIS organisation directly.

I've chosen to go with a ODS code column rather than something more
structured (perhaps a foreign key to the organisations or the locations)
because we won't always have a reference, for example if a vaccination
took place in a hospital we don't have those in our locations table
currently.

In the future we might want to make the data more structured,
particularly if we introduce a foreign key on locations to replace the
`location_name` column.