Don't create sessions when importing
When importing vaccination records we no longer need to create sessions
if they don't already exist as vaccination records are no longer tied to
a session.

Instead, for current academic years we can validate that the session
exists, and for historical imports we simply use the name of the
location (either school or clinic) to populate the `location_name`
field on the record.