Attach gillick assessments to programmes
This updates various places in the code where we handle Gillick
assessments to include the fact that they are now associated with a
programme. For now, as with everything else, this shows only the first
programme associated with a session but we expect this will be changed
once the designs are finalised.