Refactor class import routes
This removes the session slug from the class import routes to be able to
support importing class lists outside of the session page, with the
first question being the session we want to do an import for.

This has no impact on functionality but it makes it easier to build this
flow in the future and helps with building a year group selection
screen.