Add year group selection when importing class lists
This is a new feature that allows the nurses to choose the year groups
they are going to import in the class list. Depending on the year groups
selected, patients not in the year group won't be removed from the
session.

For now this is behind a feature flag (`class_import_year_groups`) so we
can deploy the code sooner and only enable the feature once it's ready
to be tested.