Update code to handle imports without programmes
Previously the cohort and immunisation import models would be attached
to a programme. This has been changed to allow imports across multiple
programmes to happen at the same time. To support this, we need to
update various places in the code that relies on the programme.

In the future we will introduce a new design for the imports which moves
them out of the programme view and in to the global view. This PR helps
us build that in the future but for now leaves the UI the same as it is
currently to allow this to be deployed first.