Filter vaccination records by programme
This follows up on #2939 to ensure that vaccination records are filtered
by the current programme. At the moment this is the first programme
administered in a session but in the future this will come from a
selection by the user.

At the same time, this also fixes a bug where vaccination records
administered outside of the current session weren't shown on a patient's
record. Now we show all vaccination records for a particular patient.