Search globally for patients when merging
This prevents a bug where a user cannot be merged with a patient that
has been moved out of the cohort as they can't be found by the policy
scope.

To support this, we've slightly updated the logic around the patient
merging to move the patient back in to the cohort if they are not
currently part of any cohort as this is likely to be the expected
behaviour.

There is an outstanding question about what happens if we're merging two
patients from two different organisation's cohorts, it's unclear which
organisation's cohort the merged patient should belong to.