Add patient to sessions when importing vaccinations
This fixes a regression that was added in
a01aa7f5632344c45f0d517d3184bd9466faa7d2 which meant that when importing
vaccination records for the current academic year (where a session must
exist) we weren't adding the patients to that session if they weren't
already in the session.

Although the data model allows patients to be vaccinated and doesn't
require them to be in the session at that point in time, in this case
it's important that the patients are visible in the session so their
status is represented accurately.