Add FullNameFormatter
This adds a class which formats a full name (made up of a given name and
a family name) in given format context, either internal use or
communication with parents.

The internal use format will use "FAMILY, Given" as the format and for
parents it will remain as "Given Family".