Show paper consent form links for each programme
This updates where we're showing the links to the paper consent forms to
show a link for each programme that the session will administer. This
doesn't quite match the designs in the prototype as the layout of the
page has changed, but it applies the change to our existing layout.