Use programme to generate patient session status
When generating the status of a patient session, we need the programme
to be available so we can generate the status approriate for the
specific programme. This updates the various parts of the code where the
status is used to ensure a programme is available.

This reduces the number of times we rely on `session.programmes.first`
and helps us to build the redesigned session pages where the programme
is selected at the top of the page.