Add ability to switch between programmes
This updates the secondary navigation on the patient session page to
show the programmes and allow for switching between the various
programmes that are administered in a particular session.

The design doesn't fully match that of the prototype, but it allows us
to start testing the basic functionality with improvements to the design
coming later.