Always add patients to community clinic
This updates how patients are added to sessions and community clinics to
always add patients to the community clinic session, even if they are
expected to be vaccinated in a school.

This allows the nurses to see all the patients in the community clinic
and perform vaccinations for them if they end up at a clinic.

The proposed solution here is a stop-gap until we handle bookings
ourselves, at which point the patients will be added to the right clinic
at the necessary time.