Invalidate self-consents on a schedule
This adds a new job which handles the logic related to invalidating
self-consents after the day of the session. Currently this behaviour is
triggered when a session is closed, but that isn't quite right.
Self-consent is only valid in the particular session date that it was
captured in, and it becomes invalid the day after.

Right now, self-consent transfers between dates in a single session and
this is not correct. As we're removing the functionality to close a
session, we need a different way of invalidating these self-consents,
and putting it in a scheduled job seemed like the quickest way to
achieve that, and it resolves the issue described above.

There are a number of problems with this approach that I would want to
fix in the future, but these problems exist with the current solution so
we're not introducing new problems:

- Gillick assessments remain valid. This is incorrect, they should also
  be invalidated, but we don't have a mechanism to do that currently.
- We invalidate triage at the same time which works fine, unless both
  the parent consented and the child consented, in which case the triage
  might still be valid as it came from the parental consent.