Don't remove patients from sessions when invalidating
Since the first release we have heard from SAIS teams that it would
actually be better if they stayed in the session so they could be
vaccinated even if their NHS number is invalid.

This also means we no longer need to add patients back in to the
sessions if they are validated after being made invalid.