Remove ability to close a session
We're removing this functionality because we find that SAIS
organisations need to close and re-open sessions to add more dates after
patients have been invited to the clinic.

This functionality wasn't doing anything because patients were already
in the clinic, and we will be replacing the clinic invitations process
with a new one to support that workflow.