Add links to consent form start page
This updates the content from the consent form journey to match the
prototype by adding the missing paragraphs with links.

`html_safe` is a little bit controversial here, but given that this is
designed to be static content and shouldn't get variables passed into it
from the `t` helper, it should be okay.