Add MenACWY and Td/IPV health questions
This adds the health questions for the MenACWY and Td/IPV vaccines and
updates the factories to be a realistic representation of the health
questions.

Like HPV, the health questions are hard coded in the vaccines seed Rake
task, but I expect in the future we'll want these to be more
customisable.