Order programmes by name
When displaying the programmes for a patient or a session we want to
ensure they appear in the same order very time. At the moment the order
is not guaranteed as we have a join table with no other reference
points.

To add this feature, I've reworked the join table to be an explicit
`SessionProgramme` model, like the `OrganisationProgramme` model and
that allow for an order to be specified on the `has_many` relationship.