Send session notifications for all programmes
If a patient is part of a programme and they're being administered a
vaccine, we should send the session reminder email to all the parents
who consents for any of the programmes in the session. In most cases,
this will be only the two parents of the child, but it's possible for a
situation where one parent consented to one programme and another parent
consented to the other.