Simplify seeds generation
This simplifies how the seeds are generated to avoid a potential never
ending loop, by only creating locations when we're setting up the
sessions. This does mean that we can no longer generate the seeds from
real school data, but we generally weren't doing that anyway since
importing all the schools takes a relatively long time and it's useful
that the seeds run quickly.

For the QA environment, we should be moving towards using the onboarding
process anyway to ensure consistency, whereas the seeds are useful for
quickly setting up a development environment.