Fetch programmes from patient session
Where we display a list of programmes, or where we select a single
programme, we now ensure that we're only selecting programmes that the
patient is eligible for, by selecting from the patient session rather
than the session.

This is possible by removing the `has_many :programmes` on the patient
session, and instead fetching all the programmes for a particular
session and then filtering by eligible year groups. We can be confident
that this change has been applied across the service as we're forced to
update the strict loading rules for everything.

In the future, this approach might be reworked such that patients belong
to multiple cohorts and we fetch the programmes via that model. This is
necessary to make it possible to easily remove patients from a cohort
even if we don't have the vaccination records for them.