Ensure consent requests are only send to eligible patients
This is necessary for the MenACWY and Td/IPV programmes which aren't
administered to Year 8s, however the HPV programme is administered for
Year 8s.

As all three vaccines might be administered at the same time as a single
session, there might be Year 8 patients in a session, and we have to
make sure that only consent requests for HPV are sent out for these
patients, not ones for Td/IPV and MenACWY as well.