Add historical_vaccination_triage_needed status
This adds a new status to patient sessions which is used to represent
patients with historical vaccination records from before they started
school. This requires the patient to go through triage before consent
requests are sent out to confirm whether the patient needs to be
vaccinated in school or not.