Allow triage without consent
This updates the triage journey to allow triaging a patient before a
consent response has been received. Effectively this is replicating
behaviour we want where patients can be removed from individual
programme cohorts, but currently this is not possible to do easily
without data model.

At the moment, all patients are considered part of a single cohort that
means if they are removed from the cohort they will be removed across
all programmes, whereas we want the ability to remove patients from
individual programmes. Since that's not possible, instead we can allow
the nurses to triage them as "do not vaccinate", preventing any consent
requests going out.