Rework how triage works for incomplete vaccination history
This changes how triage work when handling a situation where a patient
has an incomplete vaccination history to require the patient to be
triaged, even if the consent form came back as not needing triage based
on the responses for the health questions.

The other parts of the triage journey will work the same as any other
triage situation, meaning that triage has to be performed before a
patient can be vaccinated.