Add vaccine to vaccination records
At the moment the vaccine for a vaccination record comes from the batch,
but we need to be able to support uploading historical vaccination
records where we don't necessarily know the batch, but we do know the
vaccine. To record these we need to be able to store the vaccine on the
vaccination record separately to the batch.