Add SESSION_ID column when importing vaccination records
This adds a new column used when importing offline vaccination records,
to tell Mavis which session the offline records are being imported for,
ensuring that we associate them within Mavis to the correct session.

We can use the presence of this value to know if the file is being
uploaded from an offline recording and therefore the validation is
stricter than in other cases.

This is necessary because we want to be able to import vaccination
records in the current academic year that happened outside Mavis, and
currently we use that to determine whether the upload is from offline
recording or not.