Prevent removing programmes once added
Once a programme has been added to a session, it shouldn't be possible
to remove the programme, to prevent a clinic risk where a programme is
removed leading to double vaccination.

Although no data is lost when removing programmes, the data will stop
appearing in the UI and in reports.