Add programme_ids to notify log entries
This adds a new column on the notify log entries table allowing us to
track the programmes that were associated with the email or text
message.

I think this is better to store in an array to avoid the need to join
on another table when displaying the log entries.