Show notify log entry programmes in activity log
When rendering the emails and text messages that have been sent for a
particular patient, we can now show the programmes associated with that
email or text message.

To implement this I've created a cache of programmes which can be used
across all the activity log entries to avoid needing multiple queries
each time and for each type of record.