Refactor session factory
This refactors the session factory to accept only a list of programmes rather
than just a single one. There's no change to the functionality here, but
as Mavis now supports multiple programmes I think it makes sense for the
factories to follow this approach as well to encourage testing with
multiple programmes by default.