Add pagination to register attendances page
This is the only page left in the sessions view where we don't
paginate the list of patients. In most cases the list won't be too long,
but for the community clinic it could be very long and therefore might
require pagination.

I've picked a page limit of 500 because we know that the Hertfordshire
SAIS organisation is using this page to print a list of all patients in
a session, and their current largest session has 351 patients in.