Ensure validation against SESSION_ID works
Currently this was failing as the `session` method was raising an
exception if the `SESSION_ID` was present but didn't match one of the
session IDs of the organisation. Instead we can refactor this method to
return `nil` and then it can be used in other validations, specifically
when checking that the date of vaccination matches one of the dates in
the session.