Include historical records in offline reporting
This is important to include as a record of previous vaccinations so
that the nurses have all the information they need in a context where
access to Mavis is unavailable due to a lack off Internet connection.

The `SESSION_ID` will not be provided for these, which allows them to be
re-imported successfully, and there's no risk of the records being
duplicated as we rely on the `UUID` column to uniquely reference the
records.