Fix back link on parental consent confirm page
The back link wasn't working correctly, it was taking the user to the
`consent` question rather than taking the user to the last step in the
wizard flow.

There's also the potential for the previous question to have been a
health question, in which case we need to take the user to the correct
one.