Update patient table design
This is the first part of updating the design of the patient table to
match the latest designs in the prototype. In this commit, the filters
on the table are moved to the left and rendered in a feature card. On
the right the table is mostly the same, but will eventually be replaced
with a list of cards.