Add PatientSession::Record
This adds a new class that encapsulates all the logic related to patient
session vaccination outcome and statuses.

This class deals with only vaccination records that were performed in
the session, and not historical vaccination records from previously, as
it introduces a new concept of a session outcome compared to a programme
outcome.

This follows a similar approach that was added in
db31a0b76c749d3adfcaef14fb997ff21e7b7c5c.