Ensure a SchoolMove is always applied to new patients
When creating a patient for the first time during importing, we need to
make sure the `SchoolMove` is created so that the patient is attached to
the organisation, even if the school move is to "unknown school" (which
the patient would already be classed as).

This is because the school move is what adds the patient in to the
cohort, and without it, the patients will be imported but not added to
the cohort (or any sessions in the clinic).