Pre-select attendance when editing
When editing the attendance of a patient, we want the current attendance
state to be pre-selected from the radio buttons. This is to improve
usability but also fixes a bug where submitting the form with no items
selected raises an error.

The reason this wasn't happening currently is that we store the state as
`true`, `false` or `nil`, but the `nil` value can't be represented in
the radio button as it's a blank value.