Fix could not vaccinate outcome
This outcome status should only be used when it's not possible to
continue attempting to vaccinate the patient. This is only the case
where the parents have refused consent, or a nurse has triaged the
patient as unsafe to vaccinate, or a vaccination record exists with a
reason that's not retryable.