Standardise outcome class interface
This standardises the interface of the outcome classes to return a hash
object with the programmes as the key. This improves the consistency of
the classes and helps us with a future refactor where the consent,
triage and outcome classes will be moved to the patient level.

This refactor also allows us to easily memoize the values so they can be
fetched multiple times while rendering the same page.

A future design of the patient session page shows the status of all the
programmes in one page, and therefore having this available by programme
is convenient.