Move consent, triage and programme outcomes
This moves the outcomes to rely on only a patient rather than a patient
session object. This can be done without changing the data model as the
consent, triage and vaccination records are attached to the patient, not
the patient session.

This is necessary to build the programme's children tab which needs to
fetch the programme outcome for a patient regardless of the session.