Update session record and outcome tabs
This updates the tabs on the session page to match the latest designs in
the prototype. The record tab has been renamed to "Record vaccinations"
and now shows only patients that have been registered as attending, and
don't have a vaccinated session outcome. The outcome tab has been
renamed to "Session outcomes" and shows the outcome of the patients in
the session, the overall programme outcome will be displayed in the
programme children tab.