Update vaccinated criteria for Td/IPV
This updates the vaccinated criteria for Td/IPV to allow the patient to
be considered vaccinated if they have an administered vaccination record
that was recorded by a SAIS nurse in a school or community clinic
session.

The criteria in the functional spec is written as:

If a child has a Td/IPV vaccination record which is administered after the age of 10 and:
a) recorded as a fifth dose / second booster, OR
b) administered by a SAIS team
then they will be considered already vaccinated for the Td/IPV programme in Mavis and will not be added to any future sessions.