Restrict "Record vaccination" tab patients
This ensures that only patients who can be vaccinated (i.e. have consent
and triaged as safe to vaccinate) are shown in the "Record vaccination"
list of patients.

Patients who have been registered, but are missing consent or need
triage, won't appear in this list, but instead will be shown on the
"Register" tab with an action next to them.