Seed health questions when consent is given
This allows us to ensure that we're only seeding the health questions
for the relevant vaccines, for example if a parent only consents to one
vaccine then we will only show the health questions relevant for that
vaccine.

This attempts to save any previously submitted answers, if the question
title is the same.