Make fuzzy search work on full name
This should improve the reliability and predictability of the search by
comparing the similarity of the patient name as displayed to the user,
rather than comparing only the first and last name separately. In most
cases, nurses will search using the full name of the patient and the
fuzzy search is designed to avoid typos or small errors from making it
difficult to find a patient.