Set location_name to unknown when recording already had vaccine
When recording that a patient has already had the vaccine, we should set
the `location_name` to unknown where necessary (specifically this is
when we're recording the record against a clinic).

Our data model currently requires a location to be specified if
recording a vaccination record against a clinic session, or if the
session doesn't exist. This approach avoids needing to change the data
model so late in the testing process of 2.1, but I think a future
enhancement would be to allow the location column to be `nil` in certain
scenarios. Eventually we should replace the `location_name` column with
a `location_id` column anyway.