Remove `VaccinationRecord#retryable_reason?`
Although this method was being used, in both places it was irrelevant.

When determining the outcome of the programme, it had no effect as it
was being called like this (`it.not_administered? &&
!it.retryable_reason?`). The only `not_administered?` reason that's not
retryable was `already_had`, and in this case the method isn't even
called as the outcome of the programme will be vaccinated.

When determining whether the patient is safe to vaccinate has a similar
situation, at the top of the method we first check if the programme
outcome is vaccinated. In that case, either through an administered or
an `already_had` outcome, all the other reasons are retryable.