Change default TIME_OF_VACCINATION to midnight
If a `TIME_OF_VACCINATION` is not specified, the default is set to
midday (`12:00`). This causes a problem if the file is uploaded in the
morning for sessions that happened that day (which could occur for an
offline recording), as the `performed_at` date/time ends up being in
the future which fails the validation on the `VaccinationRecord`.

Setting the time to midnight ensures that this will always pass those
validation checks.

Preferrably, I think we should treat the date and time separately, and
allow for the scenario where the time of the vaccination is unknown and
present that to the user rather than needing to pick a default. However,
this would require changes to the data model.

Alternatively, my first implementation was to use the current time,
which passes the validation checks, but introduces a new issue related
to our duplicate checker. If the same file (without time of
vaccinations) is uploaded twice, because the time has now changed, the
records don't end up being flagged as duplicates, which runs the risk of
duplicate vaccination records being imported. Again, I think if we
treated the date separate to the time, we could improve this situation.