Don't send vaccination confirmations outside of session
If a vaccination record exists that was recorded outside of a school or
clinic session (i.e. it was imported) then we shouldn't be sending
parental confirmation emails for these vaccination records.

We already have a check that prevents vaccination records outside of the
current academic year from being sent, but in some cases, a patient may
be vaccinated in the current academic year at a location outside of a
school setting and this needs to be recorded in Mavis, for example if
the patient receives a vaccination at a GP surgery.

Currently, although we try to send these emails, they're not actually
succeeding as `GovukNotifyPersonalisation` requires there to be a
`session` present, and for these vaccination records there won't be.