Don't allow dose sequence to be provided
For Td/IPV and MenACWY, the dose sequence should not be provided when a
nurse is uploading a file containing vaccination records for a session
(i.e. for offline recording).

This is because we don't currently allow the nurses to set the dose
sequence while using the service normally, and it can lead to a
confusing situation where the nurse uploads a vaccination record that's
not the fifth dose, leading to the patient not being marked as
vaccinated in the service.