Don't set vaccine on vaccination record factory
It's a valid state for a vaccination record to not have a vaccine. To
ensure that we're testing for this scenario we can default to this, and
only add a vaccine where we want to explicitly test for that.

When recorded for a session in the service, the vaccine should be set.