Don't fetch school if no URN is provided
This fixes a bug where we would end up looking for a location with a
`nil` `urn` when importing a vaccination record. We weren't properly
filtering the locations meaning we would end up finding a community
clinic (since they have no URNs) and assigning the location to the name
of the community clinic.