Ensure clinic records are included in school session offline exports
When downloading the offline export for a school session, we include all
vaccinations records for the patient to ensure the nurses can see all
details about the patients. This means that some of those vaccination
records may have come from a clinic, therefore when downloading the
school spreadsheet we need to include `CLINIC_NAME` as a column to
handle those records.