Filter reports by programme
When requesting a report, we should only be returning vaccination
records for the programme the user requested the report for. At the
moment, the programme is being ignored in a number of the reports and
we're instead returning all the vaccination records for the
organisation.

I've also refactored the code to make it clearer where the vaccination
records are coming from.