Add DEPLOY_ENV, replacing SENTRY_ENVIRONMENT
We plan to start sending events to Splunk and will need to be able to
differentiate between events that go to `test` and those to `qa`. This
is a similar purpose as the SENTRY_ENVIRONMENT, so I've proposed we call
this "DEPLOY_ENV" instead defaults to the same value as "RAILS_ENV".