Don't invalidate self-consent if vaccinated
If the patient was vaccinated after getting self-consent, we shouldn't
invalidate that consent as it was valid when the vaccination was given.
A future enhancement might be to add a concept of "lifespan" to
consents, to capture exactly at which point they are invalid.

At the moment, the nurses see a patient that was vaccinated, with
supposedly no valid consent, which is not accurate.