Hide "Record as already vaccinated" without permission
Only certain users are allowed to record a vaccination, and currently we
show this link to all users. This means admins can potentially click on
this link.

Although they won't be able to complete the form (as they don't have
permission to record a vaccination), they shouldn't be shown this link
in the first place.