Set session outcome to absent from session
If the patient's attendance has been marked as absent, the session
outcome for that day should show as "absent from session".

Currently, the only way for a patient outcome to be "absent from
session" would be if the vaccination was recorded as not administered
for that reason, however we don't present "Absent from session" as one
of the options available to the nurses. Instead, the expectation is that
they would use the registration status.

In theory, you could import a vaccination record with this status, so we
still need to support that, but it's not the usual way this would work.

Potentially we could remove "absent from school" and "absent from
session" as vaccination record outcomes.

To fix this, there was a circular dependency between the
`RegisterOutcome` and `SessionOutcome` class. However, this highlighted
another issue were a patient's registeration outcome would be set to
"Session complete" if the parents refused consent. This is incorrect, as
the patient should be able to turn up on the day and self-consent.