Refactor ProgrammeOutcome, TriageOutcome and VaccinatedCriteria
This refactors a number of classes to accept a scope of patients
and determine the outcome of the patients in one go by running a
query that plucks only the necessary data and avoids the need to
instanciate lots of ActiveRecord model objects.

Ideally the three classes would be refactored in separate commits, but
they are closely tied together where changing one requires all of them
to be changed.