Call StatusUpdater
This ensures that at each point where a consent record is created or
updated, we update the consent status of the patient record.

I've decided to avoid using callbacks as I find they can lead to
suprising behaviour and this helps to keep the code explicit, however
I'm not against using callbacks if that's preferred.

This ensures that patients always have the necessary status models when
a programme is added to a session or a patient is added to a session.