Avoid preloading patient sessions
Some sessions can have lots of patient sessions (community clinics can
sometimes have upwards of 50,000 patients in) and preloading all of this
can be slow and require a large amount of memory.

Often at this scale, it seems to be more performant to not avoid N+1
queries, particularly depending on the page that is trying to be loaded.