Avoid preloading vaccination records
This avoids potential issues where preloading large amounts of data can
be slow and require a lot of memory. Instead, where N+1 issues isn't a
problem, it's fine to run an additional query to the database and fetch
only the data we need as is the case in most situations.

Now that we have cached the vaccination statuses, we can preload those
without needing to preload the entire vaccination history.