Rename disable_pds variable
This name wasn't very clear as to exactly what would be enabled/disabled
by setting the value. Instead, I've renamed the variable in Terraform
and the associated variable in Rails to make it clear that it's
referring only the jobs which update patients in bulk to reduce the
number of unnecessary PDS API calls.