Add NHS number format validation to imports
When importing historical immunisation records, we should ensure that
the NHS number is validated in the same way that it would be if a
patient was being created with that NHS number.

This has been implemented by adding a common NHS number validator which
can be used in every model that needs to validate an NHS number.