Ensure conflicting consent has correct session status
This fixes a regression that was added in the performance work where a
patient's session status was being set to "Refused" if the consent
is conflicting, which is not correct. Instead, the patient should be
left as "No outcome yet", because the consent needs to be resolved as
either given or refused.