Add BatchForm
This adds a form which handles the logic related to validating batches
when they are created or updated in the UI directly. This is needed
because we need the validation on the model to be loosened to allow for
batches without expiration dates, but when creating a batch in the UI we
want to ensure that exiration dates are always provided.